fix: track ACTUAL visual patterns, not AI explanations
System was tracking AI's text descriptions, not actual styling!

Before: Stored 'I used numbered list' (AI description)
After: Detects actual patterns: bullets, numbered_list, emojis_2, line_breaks_3, caps_1, question, short/medium/long

Visual formatter now:
1. Receives content from generator
2. Decides visual formatting (numbers, bullets, spacing, CAPS, emojis)
3. Tracks ACTUAL patterns used (not descriptions)
4. Blacklists overused patterns (e.g., if bullets used 3x, try prose)
5. Context-aware: Each generator+tone has own pattern memory

Result: No more repetitive styling (every post with bullets)

+++ src/config/config.ts
@@ -121,7 +121,7 @@ export function loadConfig(): Config {
     JOBS_LEARN_INTERVAL_MIN: process.env.JOBS_LEARN_INTERVAL_MIN ? parseInt(process.env.JOBS_LEARN_INTERVAL_MIN) : undefined,
     JOBS_POSTING_INTERVAL_MIN: process.env.JOBS_POSTING_INTERVAL_MIN ? parseInt(process.env.JOBS_POSTING_INTERVAL_MIN) : undefined,
     
-    MAX_POSTS_PER_HOUR: process.env.MAX_POSTS_PER_HOUR ? parseInt(process.env.MAX_POSTS_PER_HOUR) : undefined,
+    MAX_POSTS_PER_HOUR: process.env.MAX_POSTS_PER_HOUR ? parseFloat(process.env.MAX_POSTS_PER_HOUR) : undefined,
     REPLIES_PER_HOUR: process.env.REPLIES_PER_HOUR ? parseInt(process.env.REPLIES_PER_HOUR) : undefined,
     REPLY_MAX_PER_DAY: process.env.REPLY_MAX_PER_DAY ? parseInt(process.env.REPLY_MAX_PER_DAY) : undefined,
